+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-## 1.3.1
+ ## 1.3.1
 - added compatibility layer to support other mods that proxy
   `Minecraft.mcProfiler`, such as LiteLoader
+- implement FML coremod interface to make installation more flexible
 - added `AnimatedValue` convenience class
 - added methods to make wireframe `GLUShape`s with arbitrary line widths
 - primitive types are now serializable